feat!: new recoded canvas
BREAKING CHANGE: this breaks almost everything related to Canvas. Tui will require huge changes to gain from it.

Recode canvas to use vectorized objects to draw on it.
Objects allow to optimize canvas by having a bit of an idea of what will get drawn and allow for dynamic data to be rendered easily.

Canvas calculates intersections of an object and allows to render (or not) just part of it.
This removes neccesity to re-render every object every single frame and instead allow just to render changes.

After calculating differences and information about which rows and columns can be drawn this data gets sorted to allow creating optimized draw sequences for terminal.
This reduces stress on writing to stdout, further speeding it up.